+++ mayan/apps/documents/tests/mixins/document_version_mixins.py
@@ -71,22 +71,6 @@ class DocumentVersionAPIViewTestMixin:
             }
         )
 
-    def _request_test_document_version_export_api_view_via_get(self):
-        return self.get(
-            viewname='rest_api:documentversion-export', kwargs={
-                'document_id': self._test_document.pk,
-                'document_version_id': self._test_document.version_active.pk,
-            }
-        )
-
-    def _request_test_document_version_export_api_view_via_post(self):
-        return self.post(
-            viewname='rest_api:documentversion-export', kwargs={
-                'document_id': self._test_document.pk,
-                'document_version_id': self._test_document.version_active.pk,
-            }
-        )
-
     def _request_test_document_version_list_api_view(self):
         return self.get(
             viewname='rest_api:documentversion-list', kwargs={
@@ -286,13 +270,6 @@ class DocumentVersionViewTestMixin:
                 'document_version_id': self._test_document_version.pk
             }, data={
                 'comment': TEST_DOCUMENT_VERSION_COMMENT_EDITED
-            }
-        )
-
-    def _request_test_document_version_export_view(self):
-        return self.post(
-            viewname='documents:document_version_export', kwargs={
-                'document_version_id': self._test_document_version.pk
             }
         )